+++ data/types/bolus/factory/factory.go
@@ -6,6 +6,7 @@ import (
 	dataTypesBolusCombination "github.com/tidepool-org/platform/data/types/bolus/combination"
 	dataTypesBolusExtended "github.com/tidepool-org/platform/data/types/bolus/extended"
 	dataTypesBolusNormal "github.com/tidepool-org/platform/data/types/bolus/normal"
+	dataTypesBolusPen "github.com/tidepool-org/platform/data/types/bolus/pen"
 	"github.com/tidepool-org/platform/structure"
 	structureValidator "github.com/tidepool-org/platform/structure/validator"
 )
@@ -42,6 +43,8 @@ func NewBolusDatum(parser structure.ObjectParser) data.Datum {
 		return dataTypesBolusExtended.New()
 	case dataTypesBolusNormal.SubType:
 		return dataTypesBolusNormal.New()
+	case dataTypesBolusPen.SubType:
+		return dataTypesBolusPen.New()
 	}
 
 	parser.WithReferenceErrorReporter("subType").ReportError(structureValidator.ErrorValueStringNotOneOf(*value, subTypes))